feat: create metadata file for each release
This is an initial minimal version of the metadata JSON file, it will
be extended in the future with further fields. For now it only contains
the versions of the components, the idea is to extend it by other
information such as container hashes, potentially also container urls
and other data that is useful for automation of release tests etc.